[one-cmds] Move utils.py into onelib dir
This commit moves utils.py into onelib dir.
