+++ compiler/one-cmds/one-codegen
@@ -28,7 +28,7 @@ import os
 import sys
 import shutil
 
-import utils as _utils
+import onelib.utils as oneutils
 
 # TODO Find better way to suppress trackback on error
 sys.tracebacklimit = 0
@@ -79,7 +79,7 @@ def _get_parser(backends_list):
     parser = argparse.ArgumentParser(
         description='command line tool for code generation', usage=codegen_usage)
 
-    _utils.add_default_arg(parser)
+    oneutils.add_default_arg(parser)
 
     # get backend list in the directory
     backends_name = [ntpath.basename(f) for f in backends_list]
@@ -98,7 +98,7 @@ def _verify_arg(parser, args):
     """verify given arguments"""
     # check if required arguments is given
     missing = []
-    if not _utils.is_valid_attr(args, 'backend'):
+    if not oneutils.is_valid_attr(args, 'backend'):
         missing.append('-b/--backend')
     if len(missing):
         parser.error('the following arguments are required: ' + ' '.join(missing))
@@ -125,7 +125,7 @@ def _parse_arg(parser):
         codegen_args = parser.parse_args(codegen_args)
     # print version
     if len(args) and codegen_args.version:
-        _utils.print_version_and_exit(__file__)
+        oneutils.print_version_and_exit(__file__)
 
     return codegen_args, backend_args, unknown_args
 
@@ -139,7 +139,7 @@ def main():
     args, backend_args, unknown_args = _parse_arg(parser)
 
     # parse configuration file
-    _utils.parse_cfg(args, 'one-codegen')
+    oneutils.parse_cfg(args, 'one-codegen')
 
     # verify arguments
     _verify_arg(parser, args)
@@ -157,12 +157,12 @@ def main():
     if not codegen_path:
         raise FileNotFoundError(backend_base + ' not found')
     codegen_cmd = [codegen_path] + backend_args + unknown_args
-    if _utils.is_valid_attr(args, 'command'):
+    if oneutils.is_valid_attr(args, 'command'):
         codegen_cmd += getattr(args, 'command').split()
 
     # run backend driver
-    _utils.run(codegen_cmd, err_prefix=backend_base)
+    oneutils.run(codegen_cmd, err_prefix=backend_base)
 
 
 if __name__ == '__main__':
-    _utils.safemain(main, __file__)
+    oneutils.safemain(main, __file__)
