+++ compiler/one-cmds/one-import-bcq
@@ -25,7 +25,7 @@ import sys
 import tempfile
 
 import onelib.make_cmd as _make_cmd
-import utils as _utils
+import onelib.utils as oneutils
 import generate_bcq_output_arrays as _bcq_info_gen
 
 # TODO Find better way to suppress trackback on error
@@ -40,7 +40,7 @@ def _get_parser():
     parser = argparse.ArgumentParser(
         description='command line tool to convert TensorFlow with BCQ to circle')
 
-    _utils.add_default_arg(parser)
+    oneutils.add_default_arg(parser)
 
     ## tf2tfliteV2 arguments
     tf2tfliteV2_group = parser.add_argument_group('converter arguments')
@@ -94,9 +94,9 @@ def _verify_arg(parser, args):
     """verify given arguments"""
     # check if required arguments is given
     missing = []
-    if not _utils.is_valid_attr(args, 'input_path'):
+    if not oneutils.is_valid_attr(args, 'input_path'):
         missing.append('-i/--input_path')
-    if not _utils.is_valid_attr(args, 'output_path'):
+    if not oneutils.is_valid_attr(args, 'output_path'):
         missing.append('-o/--output_path')
     if len(missing):
         parser.error('the following arguments are required: ' + ' '.join(missing))
@@ -106,7 +106,7 @@ def _parse_arg(parser):
     args = parser.parse_args()
     # print version
     if args.version:
-        _utils.print_version_and_exit(__file__)
+        oneutils.print_version_and_exit(__file__)
 
     return args
 
@@ -115,15 +115,15 @@ def _make_generate_bcq_metadata_cmd(args, driver_path, output_path):
     """make a command for running generate_bcq_metadata"""
     cmd = [sys.executable, driver_path]
     # input_path
-    if _utils.is_valid_attr(args, 'input_path'):
+    if oneutils.is_valid_attr(args, 'input_path'):
         cmd.append('--input_path')
         cmd.append(os.path.expanduser(getattr(args, 'input_path')))
     # output_path
-    if _utils.is_valid_attr(args, 'output_path'):
+    if oneutils.is_valid_attr(args, 'output_path'):
         cmd.append('--output_path')
         cmd.append(os.path.expanduser(output_path))
     # output_arrays
-    if _utils.is_valid_attr(args, 'output_arrays'):
+    if oneutils.is_valid_attr(args, 'output_arrays'):
         cmd.append('--output_arrays')
         cmd.append(getattr(args, 'output_arrays'))
 
@@ -147,7 +147,7 @@ def _convert(args):
         f.write((' '.join(generate_bcq_metadata_cmd) + '\n').encode())
 
         # generate BCQ information metadata
-        _utils.run(generate_bcq_metadata_cmd, logfile=f)
+        oneutils.run(generate_bcq_metadata_cmd, logfile=f)
 
         # get output_arrays with BCQ
         bcq_output_arrays = _bcq_info_gen.get_bcq_output_arrays(
@@ -171,7 +171,7 @@ def _convert(args):
         f.write((' '.join(tf2tfliteV2_cmd) + '\n').encode())
 
         # convert tf to tflite
-        _utils.run(tf2tfliteV2_cmd, logfile=f)
+        oneutils.run(tf2tfliteV2_cmd, logfile=f)
 
         # make a command to convert from tflite to circle
         tflite2circle_path = os.path.join(dir_path, 'tflite2circle')
@@ -182,7 +182,7 @@ def _convert(args):
         f.write((' '.join(tflite2circle_cmd) + '\n').encode())
 
         # convert tflite to circle
-        _utils.run(tflite2circle_cmd, logfile=f)
+        oneutils.run(tflite2circle_cmd, logfile=f)
 
 
 def main():
@@ -191,7 +191,7 @@ def main():
     args = _parse_arg(parser)
 
     # parse configuration file
-    _utils.parse_cfg(args, 'one-import-bcq')
+    oneutils.parse_cfg(args, 'one-import-bcq')
 
     # verify arguments
     _verify_arg(parser, args)
@@ -201,4 +201,4 @@ def main():
 
 
 if __name__ == '__main__':
-    _utils.safemain(main, __file__)
+    oneutils.safemain(main, __file__)
