+++ compiler/one-cmds/one-import-onnx
@@ -35,7 +35,7 @@ except ImportError:
     _onnx_legalizer_enabled = False
 
 import onelib.make_cmd as _make_cmd
-import utils as _utils
+import onelib.utils as oneutils
 
 # TODO Find better way to suppress trackback on error
 sys.tracebacklimit = 0
@@ -49,7 +49,7 @@ def _get_parser():
     parser = argparse.ArgumentParser(
         description='command line tool to convert ONNX to circle')
 
-    _utils.add_default_arg(parser)
+    oneutils.add_default_arg(parser)
 
     ## tf2tfliteV2 arguments
     tf2tfliteV2_group = parser.add_argument_group('converter arguments')
@@ -105,9 +105,9 @@ def _verify_arg(parser, args):
     """verify given arguments"""
     # check if required arguments is given
     missing = []
-    if not _utils.is_valid_attr(args, 'input_path'):
+    if not oneutils.is_valid_attr(args, 'input_path'):
         missing.append('-i/--input_path')
-    if not _utils.is_valid_attr(args, 'output_path'):
+    if not oneutils.is_valid_attr(args, 'output_path'):
         missing.append('-o/--output_path')
     if len(missing):
         parser.error('the following arguments are required: ' + ' '.join(missing))
@@ -117,7 +117,7 @@ def _parse_arg(parser):
     args = parser.parse_args()
     # print version
     if args.version:
-        _utils.print_version_and_exit(__file__)
+        oneutils.print_version_and_exit(__file__)
 
     return args
 
@@ -203,18 +203,18 @@ def _convert(args):
 
     with open(logfile_path, 'wb') as f, tempfile.TemporaryDirectory() as tmpdir:
         # save intermediate
-        if _utils.is_valid_attr(args, 'save_intermediate'):
+        if oneutils.is_valid_attr(args, 'save_intermediate'):
             tmpdir = os.path.dirname(logfile_path)
         # convert onnx to tf saved model
         onnx_model = onnx.load(getattr(args, 'input_path'))
         if _onnx_legalizer_enabled:
             options = onnx_legalizer.LegalizeOptions
-            options.unroll_rnn = _utils.is_valid_attr(args, 'unroll_rnn')
-            options.unroll_lstm = _utils.is_valid_attr(args, 'unroll_lstm')
+            options.unroll_rnn = oneutils.is_valid_attr(args, 'unroll_rnn')
+            options.unroll_lstm = oneutils.is_valid_attr(args, 'unroll_lstm')
             onnx_legalizer.legalize(onnx_model, options)
-        if _utils.is_valid_attr(args, 'keep_io_order'):
+        if oneutils.is_valid_attr(args, 'keep_io_order'):
             _remap_io_names(onnx_model)
-            if _utils.is_valid_attr(args, 'save_intermediate'):
+            if oneutils.is_valid_attr(args, 'save_intermediate'):
                 basename = os.path.basename(getattr(args, 'input_path'))
                 fixed_path = os.path.join(tmpdir,
                                           os.path.splitext(basename)[0] + '~.onnx')
@@ -238,7 +238,7 @@ def _convert(args):
         f.write((' '.join(tf2tfliteV2_cmd) + '\n').encode())
 
         # convert tf to tflite
-        _utils.run(tf2tfliteV2_cmd, logfile=f)
+        oneutils.run(tf2tfliteV2_cmd, logfile=f)
 
         # make a command to convert from tflite to circle
         tflite2circle_path = os.path.join(dir_path, 'tflite2circle')
@@ -249,7 +249,7 @@ def _convert(args):
         f.write((' '.join(tflite2circle_cmd) + '\n').encode())
 
         # convert tflite to circle
-        _utils.run(tflite2circle_cmd, err_prefix="tflite2circle", logfile=f)
+        oneutils.run(tflite2circle_cmd, err_prefix="tflite2circle", logfile=f)
 
 
 def main():
@@ -258,7 +258,7 @@ def main():
     args = _parse_arg(parser)
 
     # parse configuration file
-    _utils.parse_cfg(args, 'one-import-onnx')
+    oneutils.parse_cfg(args, 'one-import-onnx')
 
     # verify arguments
     _verify_arg(parser, args)
@@ -268,4 +268,4 @@ def main():
 
 
 if __name__ == '__main__':
-    _utils.safemain(main, __file__)
+    oneutils.safemain(main, __file__)
