+++ compiler/one-cmds/one-import-tflite
@@ -24,7 +24,7 @@ import os
 import sys
 
 import onelib.make_cmd as _make_cmd
-import utils as _utils
+import onelib.utils as oneutils
 
 # TODO Find better way to suppress trackback on error
 sys.tracebacklimit = 0
@@ -38,7 +38,7 @@ def _get_parser():
     parser = argparse.ArgumentParser(
         description='command line tool to convert TensorFlow lite to circle')
 
-    _utils.add_default_arg(parser)
+    oneutils.add_default_arg(parser)
 
     ## tflite2circle arguments
     tflite2circle_group = parser.add_argument_group('converter arguments')
@@ -56,9 +56,9 @@ def _verify_arg(parser, args):
     """verify given arguments"""
     # check if required arguments is given
     missing = []
-    if not _utils.is_valid_attr(args, 'input_path'):
+    if not oneutils.is_valid_attr(args, 'input_path'):
         missing.append('-i/--input_path')
-    if not _utils.is_valid_attr(args, 'output_path'):
+    if not oneutils.is_valid_attr(args, 'output_path'):
         missing.append('-o/--output_path')
     if len(missing):
         parser.error('the following arguments are required: ' + ' '.join(missing))
@@ -68,7 +68,7 @@ def _parse_arg(parser):
     args = parser.parse_args()
     # print version
     if args.version:
-        _utils.print_version_and_exit(__file__)
+        oneutils.print_version_and_exit(__file__)
 
     return args
 
@@ -88,7 +88,7 @@ def _convert(args):
         f.write((' '.join(tflite2circle_cmd) + '\n').encode())
 
         # convert tflite to circle
-        _utils.run(tflite2circle_cmd, err_prefix="tflite2circle", logfile=f)
+        oneutils.run(tflite2circle_cmd, err_prefix="tflite2circle", logfile=f)
 
 
 def main():
@@ -97,7 +97,7 @@ def main():
     args = _parse_arg(parser)
 
     # parse configuration file
-    _utils.parse_cfg(args, 'one-import-tflite')
+    oneutils.parse_cfg(args, 'one-import-tflite')
 
     # verify arguments
     _verify_arg(parser, args)
@@ -107,4 +107,4 @@ def main():
 
 
 if __name__ == '__main__':
-    _utils.safemain(main, __file__)
+    oneutils.safemain(main, __file__)
