+++ compiler/one-cmds/one-profile
@@ -27,7 +27,7 @@ import ntpath
 import os
 import sys
 
-import utils as _utils
+import onelib.utils as oneutils
 
 # TODO Find better way to suppress trackback on error
 sys.tracebacklimit = 0
@@ -77,7 +77,7 @@ def _get_parser(backends_list):
     parser = argparse.ArgumentParser(
         description='command line tool for profiling backend model', usage=profile_usage)
 
-    _utils.add_default_arg(parser)
+    oneutils.add_default_arg(parser)
 
     # get backend list in the directory
     backends_name = [ntpath.basename(f) for f in backends_list]
@@ -96,7 +96,7 @@ def _verify_arg(parser, args):
     """verify given arguments"""
     # check if required arguments is given
     missing = []
-    if not _utils.is_valid_attr(args, 'backend'):
+    if not oneutils.is_valid_attr(args, 'backend'):
         missing.append('-b/--backend')
     if len(missing):
         parser.error('the following arguments are required: ' + ' '.join(missing))
@@ -123,7 +123,7 @@ def _parse_arg(parser):
         profile_args = parser.parse_args(profile_args)
     # print version
     if len(args) and profile_args.version:
-        _utils.print_version_and_exit(__file__)
+        oneutils.print_version_and_exit(__file__)
 
     return profile_args, backend_args, unknown_args
 
@@ -137,7 +137,7 @@ def main():
     args, backend_args, unknown_args = _parse_arg(parser)
 
     # parse configuration file
-    _utils.parse_cfg(args, 'one-profile')
+    oneutils.parse_cfg(args, 'one-profile')
 
     # verify arguments
     _verify_arg(parser, args)
@@ -151,11 +151,11 @@ def main():
     if not profile_path:
         raise FileNotFoundError(backend_base + ' not found')
     profile_cmd = [profile_path] + backend_args + unknown_args
-    if _utils.is_valid_attr(args, 'command'):
+    if oneutils.is_valid_attr(args, 'command'):
         profile_cmd += getattr(args, 'command').split()
 
     # run backend driver
-    _utils.run(profile_cmd, err_prefix=backend_base)
+    oneutils.run(profile_cmd, err_prefix=backend_base)
 
 
 if __name__ == '__main__':
