+++ compiler/one-cmds/onelib/Command.py
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import utils as oneutils
+import onelib.utils as oneutils
 
 
 class Command:
